Created qbin parser module
Qbin is the binary file format that will be used to store programs
for the QVM.
Created a generic BitInputStream that allows us to read an arbitrary
number of bits from a file.
Created container classes for sections of the qbin file.
Fully tested BitInputStream
Fully tested containers for functions.