Implemented new collections module
The collections module has been rewritten to make more sense,
and the module has been implemented but not tested.

The new classes are as follows:
CollapsingQueue: A queue which guarantees the ability to empty.
                 Right now it's just a wrapper around
                 std.container.array.Array

IdentifierMap: A data structure maintaining the list of indices and
               their types.

Program: A basic container for the Quantum Program. Note that this
         class does not contain any execution logic, but provides
         the VM with certain functions that allow it to control
         the execution flow of the quantum program.